extract IntUiBridgeButton
