+++ ide-laf-bridge/src/main/kotlin/org/jetbrains/jewel/bridge/theme/IntUiBridge.kt
@@ -239,108 +239,6 @@ internal fun createBridgeComponentStyling(theme: ThemeDefinition): ComponentStyl
     )
 }
 
-private fun readDefaultButtonStyle(): ButtonStyle {
-    val normalBackground =
-        listOf(
-                JBUI.CurrentTheme.Button.defaultButtonColorStart().toComposeColor(),
-                JBUI.CurrentTheme.Button.defaultButtonColorEnd().toComposeColor(),
-            )
-            .createVerticalBrush()
-
-    val normalContent = retrieveColorOrUnspecified("Button.default.foreground")
-
-    val normalBorder =
-        listOf(
-                JBUI.CurrentTheme.Button.buttonOutlineColorStart(true).toComposeColor(),
-                JBUI.CurrentTheme.Button.buttonOutlineColorEnd(true).toComposeColor(),
-            )
-            .createVerticalBrush()
-
-    val colors =
-        ButtonColors(
-            background = normalBackground,
-            backgroundDisabled = SolidColor(Color.Transparent),
-            backgroundFocused = normalBackground,
-            backgroundPressed = normalBackground,
-            backgroundHovered = normalBackground,
-            content = normalContent,
-            contentDisabled = retrieveColorOrUnspecified("Button.disabledText"),
-            contentFocused = normalContent,
-            contentPressed = normalContent,
-            contentHovered = normalContent,
-            border = normalBorder,
-            borderDisabled = SolidColor(JBUI.CurrentTheme.Button.disabledOutlineColor().toComposeColor()),
-            borderFocused = SolidColor(retrieveColorOrUnspecified("Button.default.focusedBorderColor")),
-            borderPressed = normalBorder,
-            borderHovered = normalBorder,
-        )
-
-    val minimumSize = JBUI.CurrentTheme.Button.minimumSize().toDpSize()
-    return ButtonStyle(
-        colors = colors,
-        metrics =
-            ButtonMetrics(
-                cornerSize = retrieveArcAsCornerSizeWithFallbacks("Button.default.arc", "Button.arc"),
-                padding = PaddingValues(horizontal = 14.dp), // see DarculaButtonUI.HORIZONTAL_PADDING
-                minSize = DpSize(minimumSize.width, minimumSize.height),
-                borderWidth = 1.dp,
-                focusOutlineExpand = 1.5.dp, // From DarculaButtonPainter.getBorderInsets
-            ),
-        focusOutlineAlignment = Stroke.Alignment.Center,
-    )
-}
-
-private fun readOutlinedButtonStyle(): ButtonStyle {
-    val normalBackground =
-        listOf(
-                JBUI.CurrentTheme.Button.buttonColorStart().toComposeColor(),
-                JBUI.CurrentTheme.Button.buttonColorEnd().toComposeColor(),
-            )
-            .createVerticalBrush()
-
-    val normalContent = retrieveColorOrUnspecified("Button.foreground")
-
-    val normalBorder =
-        listOf(
-                JBUI.CurrentTheme.Button.buttonOutlineColorStart(false).toComposeColor(),
-                JBUI.CurrentTheme.Button.buttonOutlineColorEnd(false).toComposeColor(),
-            )
-            .createVerticalBrush()
-
-    val colors =
-        ButtonColors(
-            background = normalBackground,
-            backgroundDisabled = SolidColor(Color.Transparent),
-            backgroundFocused = normalBackground,
-            backgroundPressed = normalBackground,
-            backgroundHovered = normalBackground,
-            content = normalContent,
-            contentDisabled = retrieveColorOrUnspecified("Button.disabledText"),
-            contentFocused = normalContent,
-            contentPressed = normalContent,
-            contentHovered = normalContent,
-            border = normalBorder,
-            borderDisabled = SolidColor(JBUI.CurrentTheme.Button.disabledOutlineColor().toComposeColor()),
-            borderFocused = SolidColor(JBUI.CurrentTheme.Button.focusBorderColor(false).toComposeColor()),
-            borderPressed = normalBorder,
-            borderHovered = normalBorder,
-        )
-
-    val minimumSize = JBUI.CurrentTheme.Button.minimumSize().toDpSize()
-    return ButtonStyle(
-        colors = colors,
-        metrics =
-            ButtonMetrics(
-                cornerSize = CornerSize(DarculaUIUtil.BUTTON_ARC.dp / 2),
-                padding = PaddingValues(horizontal = 14.dp), // see DarculaButtonUI.HORIZONTAL_PADDING
-                minSize = DpSize(minimumSize.width, minimumSize.height),
-                borderWidth = DarculaUIUtil.LW.dp,
-                focusOutlineExpand = Dp.Unspecified,
-            ),
-        focusOutlineAlignment = Stroke.Alignment.Center,
-    )
-}
-
 // Note: there isn't a chip spec, nor a chip UI, so we're deriving this from the
 // styling defined in com.intellij.ide.ui.experimental.meetNewUi.MeetNewUiButton
 // To note:
@@ -749,52 +647,6 @@ private fun readMenuStyle(): MenuStyle {
                 submenuMetrics = SubmenuMetrics(offset = DpOffset(0.dp, (-8).dp)),
             ),
         icons = MenuIcons(submenuChevron = AllIconsKeys.General.ChevronRight),
-    )
-}
-
-private fun readSegmentedControlButtonStyle(): SegmentedControlButtonStyle {
-    val selectedBackground = SolidColor(JBUI.CurrentTheme.SegmentedButton.SELECTED_BUTTON_COLOR.toComposeColor())
-
-    val normalBorder =
-        listOf(
-                JBUI.CurrentTheme.SegmentedButton.SELECTED_START_BORDER_COLOR.toComposeColor(),
-                JBUI.CurrentTheme.SegmentedButton.SELECTED_END_BORDER_COLOR.toComposeColor(),
-            )
-            .createVerticalBrush()
-
-    val selectedDisabledBorder =
-        listOf(
-                JBUI.CurrentTheme.Button.buttonOutlineColorStart(false).toComposeColor(),
-                JBUI.CurrentTheme.Button.buttonOutlineColorEnd(false).toComposeColor(),
-            )
-            .createVerticalBrush()
-
-    val colors =
-        SegmentedControlButtonColors(
-            background = SolidColor(Color.Transparent),
-            backgroundPressed = selectedBackground,
-            backgroundHovered = SolidColor(JBUI.CurrentTheme.ActionButton.hoverBackground().toComposeColor()),
-            backgroundSelected = selectedBackground,
-            backgroundSelectedFocused =
-                SolidColor(JBUI.CurrentTheme.SegmentedButton.FOCUSED_SELECTED_BUTTON_COLOR.toComposeColor()),
-            content = retrieveColorOrUnspecified("Button.foreground"),
-            contentDisabled = retrieveColorOrUnspecified("Label.disabledForeground"),
-            border = normalBorder,
-            borderSelected = normalBorder,
-            borderSelectedDisabled = selectedDisabledBorder,
-            borderSelectedFocused = SolidColor(JBUI.CurrentTheme.Button.focusBorderColor(false).toComposeColor()),
-        )
-
-    val minimumSize = JBUI.CurrentTheme.Button.minimumSize().toDpSize()
-    return SegmentedControlButtonStyle(
-        colors = colors,
-        metrics =
-            SegmentedControlButtonMetrics(
-                cornerSize = CornerSize(DarculaUIUtil.BUTTON_ARC.dp / 2),
-                segmentedButtonPadding = PaddingValues(horizontal = 14.dp),
-                minSize = DpSize(minimumSize.width, minimumSize.height),
-                borderWidth = DarculaUIUtil.LW.dp,
-            ),
     )
 }
